Move RouteKey class to Support\Models namespace
