+++ packages/framework/src/Routing/Route.php
@@ -9,10 +9,11 @@ use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Exceptions\RouteNotFoundException;
 use Hyde\Hyde;
 use Hyde\Support\Concerns\JsonSerializesArrayable;
+use Hyde\Support\Models\RouteKey;
 use Illuminate\Contracts\Support\Arrayable;
 use JsonSerializable;
-use function str_replace;
 use Stringable;
+use function str_replace;
 
 /**
  * The Route class bridges the gaps between Hyde pages and their respective compiled static webpages
